+++ benchmarks/memo-it.rb
@@ -15,8 +15,9 @@ end
 
 klass_a = LiteMemoizeInstanceCache.new
 klass_b = LiteMemoizeKlassCache.new
-klass_c = LiteMemoizeMixinCache.new
+klass_c = LiteMemoizeTableCache.new
 klass_d = LiteMemoizeAliasCache.new
+klass_e = LiteMemoizeVariableCache.new
 klass_z = MemoitCache.new
 
 Benchmark.ips do |x|
@@ -28,7 +29,7 @@ Benchmark.ips do |x|
     klass_b.randomize
   end
 
-  x.report('LM.mixin') do
+  x.report('LM.table') do
     klass_c.randomize
   end
 
@@ -38,6 +39,10 @@ Benchmark.ips do |x|
 
   x.report('Memoit') do
     klass_z.randomize
+  end
+
+  x.report('LM.variable') do
+    klass_e.randomize
   end
 
   x.compare!