+++ lib/lite/memoize/klass.rb
@@ -4,7 +4,7 @@ module Lite
   module Memoize
     module Klass
 
-      include Lite::Memoize::Mixin
+      include Lite::Memoize::Table
 
       def memoize(method_name, as: nil)
         inner_method = instance_method(method_name)
@@ -15,6 +15,10 @@ module Lite
           self.class.store[key] ||= inner_method.bind(self).call(*args)
         end
       end
+
+      # rubocop:disable Style/AccessModifierDeclarations
+      public :caller_key
+      # rubocop:enable Style/AccessModifierDeclarations
 
     end
   end